Move rewrite to hugr/replace
